Extracted test utils
