+++ tests/TestUtils/TestBusEngine.cs
@@ -53,15 +53,4 @@ namespace Tests
         }
     }
 
-
-    public class FirstTestCommand : ICommand
-    {
-        public string Message { get; set; }
-    }
-
-    public class SecondTestCommand : ICommand { }
-
-    public class FirstTestEvent : IEvent { }
-
-    public class SecondTestEvent : IEvent { }
 }
