--- tests/Tests.Integration.Nybus/TestTypes.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Nybus;
-
-namespace Tests
-{
-    public class FirstTestCommand : ICommand
-    {
-        public string Message { get; set; }
-    }
-
-    public class SecondTestCommand : ICommand { }
-
-    public class FirstTestEvent : IEvent { }
-
-    public class SecondTestEvent : IEvent { }
-}
