+++ tests/Tests.Nybus.Abstractions/Tests.Nybus.Abstractions.csproj
@@ -14,6 +14,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Nybus.Abstractions\Nybus.Abstractions.csproj" />
+    <ProjectReference Include="..\TestUtils\TestUtils.csproj" />
   </ItemGroup>
 
 </Project>
