+++ tests/Tests.Nybus.Engine.RabbitMq/Configuration/RabbitMqConfiguratorTests.cs
@@ -5,9 +5,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Moq;
 using NUnit.Framework;
-using Nybus;
 using Nybus.Configuration;
-using Nybus.Policies;
 
 namespace Tests.Configuration
 {
@@ -114,47 +112,6 @@ namespace Tests.Configuration
         public string ProviderName { get; }
 
         public IQueueFactory CreateFactory(IConfigurationSection settings)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    public class TestNybusConfigurator : INybusConfigurator
-    {
-        public void UseConfiguration(Microsoft.Extensions.Configuration.IConfiguration configuration, string sectionName = "Nybus")
-        {
-            Configuration = configuration.GetSection(sectionName);
-        }
-
-        private readonly List<Action<IServiceCollection>> _serviceConfigurations = new List<Action<IServiceCollection>>();
-
-        public void AddServiceConfiguration(Action<IServiceCollection> configurator)
-        {
-            _serviceConfigurations.Add(configurator);
-        }
-
-        public void ApplyServiceConfigurations(IServiceCollection services)
-        {
-            foreach (var sc in _serviceConfigurations)
-                sc(services);
-        }
-
-        private readonly List<Action<ISubscriptionBuilder>> _subscriptionBuilders = new List<Action<ISubscriptionBuilder>>();
-
-        public void AddSubscription(Action<ISubscriptionBuilder> configurator)
-        {
-            _subscriptionBuilders.Add(configurator);
-        }
-
-        public void ApplySubscriptions(ISubscriptionBuilder builder)
-        {
-            foreach (var sb in _subscriptionBuilders)
-                sb(builder);
-        }
-
-        public Microsoft.Extensions.Configuration.IConfiguration Configuration { get; private set; }
-
-        public void Configure(Action<INybusConfiguration> configuration)
         {
             throw new NotImplementedException();
         }
